修正 多对多连接

+++ src/main/java/top/ncserver/chatsync/Client.java
@@ -52,8 +52,8 @@ public class Client extends BukkitRunnable {
                         Map<String, Object> msg1 = new HashMap<>();
                         msg1.put("type", "init");
                         msg1.put("name", Chatsync.config.getString("ServerName"));
-                        if (Chatsync.config.getInt("GourpID")!=111){
-                            msg1.put("groupId",Chatsync.config.getInt("GourpID"));
+                        if (Chatsync.config.getLong("GourpID")!=111L){
+                            msg1.put("groupId",Chatsync.config.getLong("GourpID"));
                         }
 
                         JSONObject jo = new JSONObject(msg1);
